requirements.txt
yt-dlp
youtube-transcript-api
opencv-python
PyMuPDF
sentence-transformers
Pillow
numpy
llama-index
llama-index-embeddings-huggingface
llama-index-vector-stores-lancedb
lancedb
pyarrow
imagehash
transformers
openai-whisper
whisperx
google-generativeai
langgraph
langgraph-checkpoint-sqlite
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic-settings
slowapi
loguru
redis
celery[redis]
requests
